Add delete mechanics with button

+++ backend/graphql/mutation.js
@@ -65,8 +65,12 @@ const RootMutationType = new GraphQLObjectType({
             args: {
                 _id: { type: GraphQLNonNull(GraphQLID) }
             },
-            // TODO Add authentification check
-            resolve: async function(_, args) {
+            resolve: async function(_, args, context) {
+                if (!context.req.isAuth) {
+                    // IF not authenticated this error-text will show in graphql-response
+                    throw new Error("Not authenticated.");
+                }
+
                 try {
                     const deleteTicket = await ticketsModel.deleteTicket(args);
